Created shared dialog for mpga and staff expense reports
