+++ src/components/Reports/MPGAIncomeExpensesReport/BreakdownModal/BreakdownModal.tsx
@@ -1,9 +1,5 @@
 import { useMemo } from 'react';
 import {
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
   Table,
   TableBody,
   TableCell,
@@ -15,7 +11,6 @@ import {
 } from '@mui/material';
 import { DateTime } from 'luxon';
 import { useTranslation } from 'react-i18next';
-import { CancelButton } from 'src/components/Shared/Modal/ActionButtons/ActionButtons';
 import {
   StaffExpenseCategoryEnum,
   StaffExpensesSubCategoryEnum,
@@ -23,6 +18,7 @@ import {
 import { useLocale } from 'src/hooks/useLocale';
 import { currencyFormat } from 'src/lib/intlFormat';
 import theme from 'src/theme';
+import { DialogSkeleton } from '../../Shared/DialogSkeleton/DialogSkeleton';
 import { getLocalizedCategory } from '../../Shared/Helpers/transformStaffExpenseEnums';
 import { BreakdownAccordion } from '../BreakdownAccordion/BreakdownAccordion';
 import { TransactionBreakdown } from '../mockData';
@@ -82,108 +78,82 @@ export const BreakdownModal: React.FC<BreakdownModalProps> = ({
   );
 
   return (
-    <Dialog
+    <DialogSkeleton
+      categoryName={getLocalizedCategory(category, t)}
       open={open}
       onClose={onClose}
-      maxWidth="md"
-      fullWidth
-      PaperProps={{
-        sx: {
-          maxHeight: '700px',
-        },
-      }}
     >
-      <DialogTitle>
-        {t(`${getLocalizedCategory(category, t)} Breakdown`)}
-      </DialogTitle>
-      <DialogContent
-        sx={{
-          padding: theme.spacing(4),
-          overflow: 'hidden',
-          display: 'flex',
-        }}
-      >
-        <TableContainer>
-          <Table>
-            <TableHead>
-              <TableRow
-                sx={{ backgroundColor: theme.palette.mpdxGrayLight.main }}
-              >
-                <TableCell>{t('Category')}</TableCell>
-                <TableCell sx={{ textAlign: 'right' }}>
-                  {t(
-                    `${startDate.toFormat('LLLL yyyy')} - ${endDate.toFormat('LLLL yyyy')}`,
-                  )}
-                </TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {subcategoryBreakdown.map(
-                ({ subCategory, transactions, total }) => (
-                  <TableRow key={subCategory}>
-                    <TableCell colSpan={2} sx={{ padding: 0, border: 0 }}>
-                      <BreakdownAccordion
-                        category={category}
-                        subCategory={subCategory}
-                        transactions={transactions}
-                        total={total}
-                      />
-                    </TableCell>
-                  </TableRow>
-                ),
-              )}
-            </TableBody>
-            <TableFooter
-              sx={{
-                '& .MuiTableCell-footer': {
-                  position: 'sticky',
-                  bottom: 0,
-                  backgroundColor: 'background.paper',
-                  borderBottom: 0,
-                  boxShadow: `inset 0 -1px 0 ${theme.palette.divider}`,
-                },
-              }}
+      <TableContainer>
+        <Table>
+          <TableHead>
+            <TableRow
+              sx={{ backgroundColor: theme.palette.mpdxGrayLight.main }}
             >
-              <TableRow>
-                <TableCell>
-                  <Typography
-                    color={theme.palette.text.primary}
-                    fontWeight="bold"
-                  >
-                    {t(
-                      `Total ${getLocalizedCategory(category, t)} ${overallTotal >= 0 ? 'Income' : 'Expense'}`,
-                    )}
-                  </Typography>
-                </TableCell>
-                <TableCell align="right">
-                  <Typography
-                    fontWeight="bold"
-                    sx={{
-                      color:
-                        overallTotal >= 0
-                          ? theme.palette.statusSuccess.main
-                          : theme.palette.chipRedDark.main,
-                    }}
-                  >
-                    {currencyFormat(Math.abs(overallTotal), currency, locale, {
-                      showTrailingZeros: true,
-                    })}
-                  </Typography>
-                </TableCell>
-              </TableRow>
-            </TableFooter>
-          </Table>
-        </TableContainer>
-      </DialogContent>
-      <DialogActions
-        sx={{
-          padding: theme.spacing(2),
-        }}
-      >
-        <CancelButton data-testid="close-button" onClick={onClose}>
-          {t('Close')}
-        </CancelButton>
-      </DialogActions>
-    </Dialog>
+              <TableCell>{t('Category')}</TableCell>
+              <TableCell sx={{ textAlign: 'right' }}>
+                {t(
+                  `${startDate.toFormat('LLLL yyyy')} - ${endDate.toFormat('LLLL yyyy')}`,
+                )}
+              </TableCell>
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {subcategoryBreakdown.map(
+              ({ subCategory, transactions, total }) => (
+                <TableRow key={subCategory}>
+                  <TableCell colSpan={2} sx={{ padding: 0, border: 0 }}>
+                    <BreakdownAccordion
+                      category={category}
+                      subCategory={subCategory}
+                      transactions={transactions}
+                      total={total}
+                    />
+                  </TableCell>
+                </TableRow>
+              ),
+            )}
+          </TableBody>
+          <TableFooter
+            sx={{
+              '& .MuiTableCell-footer': {
+                position: 'sticky',
+                bottom: 0,
+                backgroundColor: 'background.paper',
+                borderBottom: 0,
+                boxShadow: `inset 0 -1px 0 ${theme.palette.divider}`,
+              },
+            }}
+          >
+            <TableRow>
+              <TableCell>
+                <Typography
+                  color={theme.palette.text.primary}
+                  fontWeight="bold"
+                >
+                  {t(
+                    `Total ${getLocalizedCategory(category, t)} ${overallTotal >= 0 ? 'Income' : 'Expense'}`,
+                  )}
+                </Typography>
+              </TableCell>
+              <TableCell align="right">
+                <Typography
+                  fontWeight="bold"
+                  sx={{
+                    color:
+                      overallTotal >= 0
+                        ? theme.palette.statusSuccess.main
+                        : theme.palette.chipRedDark.main,
+                  }}
+                >
+                  {currencyFormat(Math.abs(overallTotal), currency, locale, {
+                    showTrailingZeros: true,
+                  })}
+                </Typography>
+              </TableCell>
+            </TableRow>
+          </TableFooter>
+        </Table>
+      </TableContainer>
+    </DialogSkeleton>
   );
 };
