+++ src/components/Reports/StaffExpenseReport/CategoryBreakdownDialog/CategoryBreakdownDialog.tsx
@@ -1,9 +1,5 @@
 import { useMemo } from 'react';
 import {
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
   Table,
   TableBody,
   TableCell,
@@ -16,9 +12,9 @@ import {
 import { useTheme } from '@mui/material/styles';
 import { DateTime } from 'luxon';
 import { useTranslation } from 'react-i18next';
-import { CancelButton } from 'src/components/Shared/Modal/ActionButtons/ActionButtons';
 import { useLocale } from 'src/hooks/useLocale';
 import { currencyFormat, dateFormat } from 'src/lib/intlFormat';
+import { DialogSkeleton } from '../../Shared/DialogSkeleton/DialogSkeleton';
 import { Transaction } from '../Helpers/filterTransactions';
 
 export interface CategoryBreakdownDialogProps {
@@ -45,110 +41,80 @@ export const CategoryBreakdownDialog: React.FC<
   );
 
   return (
-    <Dialog
-      open={isOpen}
-      onClose={onClose}
-      maxWidth="md"
-      fullWidth
-      PaperProps={{
-        sx: {
-          maxHeight: '700px',
-        },
-      }}
-    >
-      <DialogTitle>
-        {categoryName} {t('Breakdown')}
-      </DialogTitle>
-      <DialogContent
-        sx={{
-          padding: theme.spacing(4),
-          overflow: 'hidden',
-          display: 'flex',
-        }}
-      >
-        <TableContainer>
-          <Table stickyHeader>
-            <TableHead>
-              <TableRow
-                sx={{
-                  backgroundColor: 'background.paper',
-                  fontWeight: 'bold',
-                }}
-              >
-                <TableCell width={150}>{t('Date')}</TableCell>
-                <TableCell width={200}>{t('Description')}</TableCell>
-                <TableCell width={200}>{t('Category')}</TableCell>
-                <TableCell align="right" width={150}>
-                  {t('Amount')}
-                </TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {transactionsSortedByDate.map((transaction) => (
-                <TableRow key={transaction.id}>
-                  <TableCell>
-                    {dateFormat(
-                      DateTime.fromISO(transaction.transactedAt),
-                      locale,
-                    )}
-                  </TableCell>
-                  <TableCell>{transaction.description}</TableCell>
-                  <TableCell>{transaction.displayCategory}</TableCell>
-                  <TableCell align="right">
-                    {currencyFormat(transaction.amount, 'USD', locale)}
-                  </TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-            <TableFooter
+    <DialogSkeleton categoryName={categoryName} open={isOpen} onClose={onClose}>
+      <TableContainer>
+        <Table stickyHeader>
+          <TableHead>
+            <TableRow
               sx={{
-                position: 'sticky',
-                bottom: 0,
                 backgroundColor: 'background.paper',
+                fontWeight: 'bold',
               }}
             >
-              <TableRow>
-                <TableCell colSpan={3}>
-                  <Typography
-                    color={theme.palette.text.primary}
-                    fontWeight="bold"
-                  >
-                    {totalAmount > 0
-                      ? t(`Total {{category}} Income`, {
-                          category: categoryName,
-                        })
-                      : t(`Total {{category}} Expense`, {
-                          category: categoryName,
-                        })}
-                  </Typography>
+              <TableCell width={150}>{t('Date')}</TableCell>
+              <TableCell width={200}>{t('Description')}</TableCell>
+              <TableCell width={200}>{t('Category')}</TableCell>
+              <TableCell align="right" width={150}>
+                {t('Amount')}
+              </TableCell>
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {transactionsSortedByDate.map((transaction) => (
+              <TableRow key={transaction.id}>
+                <TableCell>
+                  {dateFormat(
+                    DateTime.fromISO(transaction.transactedAt),
+                    locale,
+                  )}
                 </TableCell>
+                <TableCell>{transaction.description}</TableCell>
+                <TableCell>{transaction.displayCategory}</TableCell>
                 <TableCell align="right">
-                  <Typography
-                    fontWeight="bold"
-                    sx={{
-                      color:
-                        totalAmount > 0
-                          ? theme.palette.success.main
-                          : theme.palette.error.main,
-                    }}
-                  >
-                    {currencyFormat(totalAmount, 'USD', locale)}
-                  </Typography>
+                  {currencyFormat(transaction.amount, 'USD', locale)}
                 </TableCell>
               </TableRow>
-            </TableFooter>
-          </Table>
-        </TableContainer>
-      </DialogContent>
-      <DialogActions
-        sx={{
-          padding: theme.spacing(2),
-        }}
-      >
-        <CancelButton data-testid="close-button" onClick={onClose}>
-          {t('Close')}
-        </CancelButton>
-      </DialogActions>
-    </Dialog>
+            ))}
+          </TableBody>
+          <TableFooter
+            sx={{
+              position: 'sticky',
+              bottom: 0,
+              backgroundColor: 'background.paper',
+            }}
+          >
+            <TableRow>
+              <TableCell colSpan={3}>
+                <Typography
+                  color={theme.palette.text.primary}
+                  fontWeight="bold"
+                >
+                  {totalAmount > 0
+                    ? t(`Total {{category}} Income`, {
+                        category: categoryName,
+                      })
+                    : t(`Total {{category}} Expense`, {
+                        category: categoryName,
+                      })}
+                </Typography>
+              </TableCell>
+              <TableCell align="right">
+                <Typography
+                  fontWeight="bold"
+                  sx={{
+                    color:
+                      totalAmount > 0
+                        ? theme.palette.success.main
+                        : theme.palette.error.main,
+                  }}
+                >
+                  {currencyFormat(totalAmount, 'USD', locale)}
+                </Typography>
+              </TableCell>
+            </TableRow>
+          </TableFooter>
+        </Table>
+      </TableContainer>
+    </DialogSkeleton>
   );
 };
